Separar crear orden trab y ordenes trab
