+++ src/components/OrdenTrabajo.js
@@ -39,23 +39,6 @@ function OrdenTrabajo() {
         }
     }, [currentUser]);*/
 
-    const [ordenes, setOrdenes] = useState([]);
-
-    useEffect(() => {
-        const fetchOrders = async () => {
-            const { data, error } = await supabase
-                .from('ordentrabajo')
-                .select('*');
-
-            if (error) {
-                console.error('Error fetching data:', error);
-            } else {
-                setOrdenes(data);
-            }
-        };
-
-        fetchOrders();
-    }, []);
 
     const handleChange = (e) => {
         const { name, value, type, checked } = e.target;
@@ -71,6 +54,8 @@ function OrdenTrabajo() {
             setFormData({ ...formData, [name]: value });
         }
     };
+
+    const [ordenes, setOrdenes] = useState([]);
 
     const handleSubmit = async (e) => {
         e.preventDefault();
@@ -115,25 +100,6 @@ function OrdenTrabajo() {
                 horaRecibida: '',
                 recibidaPor: ''
             });
-        }
-    };
-   
-    const deleteOrden = async (id) => {
-        if (!id) {
-            console.error('Invalid ID provided for deletion:', id);
-            return;
-        }
-    
-        const { error } = await supabase
-            .from('ordentrabajo')
-            .delete()
-            .eq('id', id);
-    
-        if (error) {
-            console.error('Error deleting data:', error);
-        } else {
-            console.log('Order deleted successfully.');
-            setOrdenes(prevOrdenes => prevOrdenes.filter(orden => orden.id !== id));
         }
     };
     
@@ -222,33 +188,7 @@ function OrdenTrabajo() {
 
                 <button type="submit">Enviar</button>
             </form>
-            <h3>Órdenes de Trabajo</h3>
-            <table className="user-table">
-                <thead>
-                    <tr>
-                        <th>Cliente</th>
-                        <th>Teléfono</th>
-                        <th>Fecha</th>
-                        <th>Hora</th>
-                        <th>Servicios</th>
-                        <th>Acciones</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {ordenes.map((orden) => (
-                        <tr key={orden.id}>
-                            <td>{orden.cliente}</td>
-                            <td>{orden.telefono}</td>
-                            <td>{orden.fecha}</td>
-                            <td>{orden.hora}</td>
-                            <td>{orden.servicios.join(', ')}</td>
-                            <td>
-                                <button onClick={() => deleteOrden(orden.id)}>Eliminar</button>
-                            </td>
-                        </tr>
-                    ))}
-                </tbody>
-            </table>
+            
         </div>
